+++ src/anemoi/datasets/create/filters/orog_to_z.py
@@ -12,8 +12,8 @@ from typing import Any
 from typing import Dict
 
 import earthkit.data as ekd
-from anemoi.transform.field import new_field_from_numpy
-from anemoi.transform.field import new_fieldlist_from_list
+from anemoi.transform.fields import new_field_from_numpy
+from anemoi.transform.fields import new_fieldlist_from_list
 
 
 def execute(context: Any, input: ekd.FieldList, orog: str, z: str = "z") -> ekd.FieldList:
