+++ src/anemoi/datasets/create/filters/single_level_specific_humidity_to_relative_humidity.py
@@ -16,8 +16,8 @@ from typing import Union
 
 import earthkit.data as ekd
 import numpy as np
-from anemoi.transform.field import new_field_from_numpy
-from anemoi.transform.field import new_fieldlist_from_list
+from anemoi.transform.fields import new_field_from_numpy
+from anemoi.transform.fields import new_fieldlist_from_list
 from earthkit.meteo import constants
 from earthkit.meteo import thermo
 from numpy.typing import NDArray
