+++ src/anemoi/datasets/create/filters/sum.py
@@ -15,8 +15,8 @@ from typing import List
 from typing import Tuple
 
 import earthkit.data as ekd
-from anemoi.transform.field import new_field_from_numpy
-from anemoi.transform.field import new_fieldlist_from_list
+from anemoi.transform.fields import new_field_from_numpy
+from anemoi.transform.fields import new_fieldlist_from_list
 
 
 def execute(context: Any, input: ekd.FieldList, params: List[str], output: str) -> ekd.FieldList:
