+++ src/anemoi/datasets/create/filters/unrotate_winds.py
@@ -11,8 +11,8 @@ from collections import defaultdict
 from typing import Any
 
 import earthkit.data as ekd
-from anemoi.transform.field import new_field_from_numpy
-from anemoi.transform.field import new_fieldlist_from_list
+from anemoi.transform.fields import new_field_from_numpy
+from anemoi.transform.fields import new_fieldlist_from_list
 
 # import numpy as np
 from earthkit.geo.rotate import unrotate_vector
